eric6_sources.leo is almost 20 MB. It contains all .py files in eric6, that is, in the eric6/eric folder.
This is a "safe" file, using @@clean for all files.

Three files failed to import perfectly at first.  All but one now do, as explained in the outline. git diff (for those files) guarantees that my by-hand changes were accurate.

There is lots of interesting code in this outline.  I'll be studying it closely.